// src/Model/iTechRestApi/iTechControlCollectionEnum.ts
//*************************DO NOT MODIFY**************************
//
//THESE FILES ARE AUTOGENERATED WITH A TYPESCRIPT CODE GENERATION TOOL AND ANY MODIFICATIONS MADE HERE WILL BE LOST.
//PLEASE VISIT http://frhagn.github.io/Typewriter/ TO LEARN MORE ABOUT THIS VISUAL STUDIO EXTENSION
//
//*************************DO NOT MODIFY**************************



export enum iTechControlCollectionEnum {
    truphoneVoice = 1,
    fixedVoice = 2,
    skypeVoice = 3,
    skypeIM = 4,
    turretVoice = 5,
    tmlVoice = 6,
    tmlim = 7,
    truphoneSMS = 8,
    bloombergIM = 9,
    reutersIM = 10,
    symphonyIM = 11,
    eMail = 12,
    telewareRecall = 13,
    byodApp = 14,
    kpnMobile = 15,
    telemessageVoice = 16,
    telemessageIM = 17,
    tmWhatsappVoice = 18,
    tmWhatsappIM = 19,
    redboxArchive = 20,
    webEx = 21,
    teamsVoice = 22,
    teamsIM = 23,
    teamsVideo = 24
}